backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
pandas
numpy
mlxtend
statsmodels
scikit-learn
scipy
matplotlib
seaborn
werkzeug
python-dotenv
